backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
gunicorn>=21.2.0
pytest
pytest-django
